Last chanes before moving to the new clone
