+++ app/dashboard/courier/workspace.code-workspace
@@ -1,0 +1,8 @@
+{
+	"folders": [
+		{
+			"path": "../../../../Logi/logistics-enhanced"
+		}
+	],
+	"settings": {}
+}
